move the pratica child components
