+++ src/pages/core/core.module.ts
@@ -5,7 +5,6 @@ import { ComponentsModule } from '../../components/components.module';
 import { PracticesPageModule } from './practices/practices.module';
 import { TemparioPageModule } from './tempario/tempario.module';
 import { RentalPageModule } from './rental/rental.module';
-import { CoreComponentsModule } from './shared/core-compoentns.module';
 import { PdfPreviewPageModule } from './pdf-preview/pdf-preview.module';
 import { SignatureModalPageModule } from './signature-modal/signature-modal.module';
 
@@ -16,7 +15,6 @@ import { SignatureModalPageModule } from './signature-modal/signature-modal.modu
   imports: [
     IonicPageModule.forChild(CorePage),
     ComponentsModule,
-    CoreComponentsModule,
     PracticesPageModule,
     TemparioPageModule,
     RentalPageModule,
