+++ src/pages/core/practices/shared/budgets-tab/budgets-tab.scss
