+++ src/pages/core/practices/shared/documents-tab/documents-tab.scss
