+++ src/pages/core/practices/shared/documents-tab/documents-tab.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { NavController, NavParams } from 'ionic-angular';
-import { PdfPreviewPage } from '../../pdf-preview/pdf-preview';
+import { PdfPreviewPage } from '../../../pdf-preview/pdf-preview';
 
 @Component({
   selector: 'documents-tab',
