+++ src/pages/core/practices/shared/outlook-tab/outlook-tab.scss
