+++ src/pages/core/practices/shared/payment-tab/payment-tab.ts
