+++ src/pages/core/practices/shared/photo-tab/photo-tab.scss
