+++ src/pages/core/practices/shared/pratica-compoentns.module.ts
@@ -1,13 +1,13 @@
 import { NgModule } from '@angular/core';
 import { IonicModule } from "ionic-angular";
-import { ComponentsModule } from "../../../components/components.module";
+import { ComponentsModule } from "../../../../components/components.module";
 import { OutlookTabComponet } from './outlook-tab/outlook-tab';
 import { BudgetsTabComponent } from './budgets-tab/budgets-tab';
 import { DocumentsTabComponent } from './documents-tab/documents-tab';
 import { PaymentTabComponet } from './payment-tab/payment-tab';
 import { PhotoTabComponet } from './photo-tab/photo-tab';
 import { RentTabComponet } from './rent-tab/rent-tab';
-import { PdfPreviewPageModule } from '../pdf-preview/pdf-preview.module';
+import { PdfPreviewPageModule } from '../../pdf-preview/pdf-preview.module';
 import { NgxChartsModule } from '@swimlane/ngx-charts';
 
 
@@ -34,4 +34,4 @@ import { NgxChartsModule } from '@swimlane/ngx-charts';
     RentTabComponet
   ]
 })
-export class CoreComponentsModule {}
+export class PraticaComponentsModule {}
