+++ src/pages/core/practices/shared/rent-tab/rent-tab.html
@@ -64,7 +64,6 @@
           </ion-row>
         </ion-grid>
       </div>
-
       
     </div>
   </div>
