+++ src/pages/core/practices/shared/rent-tab/rent-tab.scss
